Move HIE files stuff to a new hie-compat package
