+++ ghcide.cabal
@@ -54,6 +54,7 @@ library
         hashable,
         haskell-lsp-types == 0.22.*,
         haskell-lsp == 0.22.*,
+        hie-compat,
         mtl,
         network-uri,
         prettyprinter-ansi-terminal,
@@ -184,24 +185,6 @@ library
         Development.IDE.Plugin.Completions.Logic
         Development.IDE.Plugin.Completions.Types
         Development.IDE.Types.Action
-    if (impl(ghc > 8.5) && impl(ghc < 8.7)) && !flag(ghc-lib)
-      hs-source-dirs: src-ghc86
-      other-modules:
-        Development.IDE.GHC.HieAst
-        Development.IDE.GHC.HieBin
-        Development.IDE.GHC.HieTypes
-        Development.IDE.GHC.HieDebug
-        Development.IDE.GHC.HieUtils
-    if (impl(ghc > 8.7) && impl(ghc < 8.10)) || flag(ghc-lib)
-      hs-source-dirs: src-ghc88
-      other-modules:
-        Development.IDE.GHC.HieAst
-        Development.IDE.GHC.HieBin
-    if (impl(ghc > 8.9))
-      hs-source-dirs: src-ghc810
-      other-modules:
-        Development.IDE.GHC.HieAst
-        Development.IDE.GHC.HieBin
     ghc-options: -Wall -Wno-name-shadowing -Wincomplete-uni-patterns
 
 executable ghcide-test-preprocessor
