+++ hie-compat/CHANGELOG.md
@@ -1,0 +1,5 @@
+# Revision history for hie-compat
+
+## 0.1.0.0 -- 2020-10-19
+
+* Initial Release
