+++ hie-compat/hie-compat.cabal
@@ -1,0 +1,46 @@
+cabal-version:       >=1.10
+name:                hie-compat
+version:             0.1.0.0
+synopsis:            HIE files for GHC 8.6 and other HIE file backports
+license:             Apache-2.0
+description:
+  Backports for HIE files to GHC 8.6, along with a few other backports
+  of HIE file related fixes for ghcide
+  THIS DOES NOT LET YOU READ HIE FILES WITH MISMATCHED VERSIONS OF GHC
+license-file:        LICENSE
+author:              Zubin Duggal
+maintainer:          zubin.duggal@gmail.com
+build-type:          Simple
+extra-source-files:  CHANGELOG.md
+
+library
+  build-depends:
+     ghc >= 8.6, base, array, bytestring, containers, directory, filepath,
+     ghc-boot, transformers
+  default-language:    Haskell2010
+
+  if (impl(ghc > 8.5) && impl(ghc < 8.7))
+    hs-source-dirs: src-ghc86
+    exposed-modules:
+      Compat.HieAst
+      Compat.HieBin
+      Compat.HieTypes
+      Compat.HieDebug
+      Compat.HieUtils
+  if (impl(ghc > 8.7) && impl(ghc < 8.10))
+    hs-source-dirs: src-ghc88
+    exposed-modules:
+      Compat.HieAst
+      Compat.HieBin
+  if (impl(ghc > 8.9) && impl(ghc < 8.11))
+    hs-source-dirs: src-ghc810
+    exposed-modules:
+      Compat.HieAst
+      Compat.HieBin
+
+  if (impl(ghc > 8.7) && impl(ghc < 8.11))
+    reexported-modules:
+      HieTypes as Compat.HieTypes,
+      HieDebug as Compat.HieDebug,
+      HieUtils as Compat.HieUtils
+
