+++ hie-compat/src-ghc86/Compat/HieDebug.hs
@@ -4,7 +4,7 @@ Functions to validate and check .hie file ASTs generated by GHC.
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE FlexibleContexts #-}
-module Development.IDE.GHC.HieDebug where
+module Compat.HieDebug where
 
 import Prelude hiding ((<>))
 import SrcLoc
@@ -12,9 +12,9 @@ import Module
 import FastString
 import Outputable
 
-import Development.IDE.GHC.HieTypes
-import Development.IDE.GHC.HieBin
-import Development.IDE.GHC.HieUtils
+import Compat.HieTypes
+import Compat.HieBin
+import Compat.HieUtils
 
 import qualified Data.Map as M
 import qualified Data.Set as S
