+++ stack-ghc-lib.yaml
@@ -1,6 +1,7 @@
 resolver: nightly-2019-09-16
 packages:
 - .
+- ./hie-compat/
 extra-deps:
 - haskell-lsp-0.22.0.0
 - haskell-lsp-types-0.22.0.0
