+++ stack88.yaml
@@ -1,6 +1,7 @@
 resolver: lts-16.12
 packages:
 - .
+- ./hie-compat/
 extra-deps:
 - haskell-lsp-0.22.0.0
 - haskell-lsp-types-0.22.0.0
